Refactor: 컴포넌트별 폴더 생성 및 분리작업
