+++ src/pages/ShopList/Pagination/Pagination.scss
